Add a couple escape hatches to logging config.
First, the business of getting a logger instance has been relegated to a
`Vassal.get_logger` instance method. This way, subclasses can simply
override that method to implement custom behavior.

Second, `script_main` now accepts a `_setup_logging` boolean. If it's
true (the default), the `setup_logging` function is invoked as
normal. Otherwise, it's not called. This is to support situations where
we are using something other than a file-based logging config.